+++ src/cryptography/hazmat/primitives/serialization/__init__.py
@@ -23,6 +23,7 @@ from cryptography.hazmat.primitives.serialization.base import (
 )
 from cryptography.hazmat.primitives.serialization.ssh import (
     load_ssh_private_key,
+    load_ssh_public_identity,
     load_ssh_public_key,
 )
 
@@ -34,6 +35,7 @@ __all__ = [
     "load_pem_private_key",
     "load_pem_public_key",
     "load_ssh_private_key",
+    "load_ssh_public_identity",
     "load_ssh_public_key",
     "Encoding",
     "PrivateFormat",