+++ CHANGELOG.rst
@@ -10,6 +10,10 @@ Changelog
 
 * Support for Python 3.6 is deprecated and will be removed in the next
   release.
+* Added support for parsing SSH certificates in addition to public keys with
+  :func:`~cryptography.hazmat.primitives.serialization.load_ssh_public_identity`.
+  :func:`~cryptography.hazmat.primitives.serialization.load_ssh_public_key`
+  continues to support only public keys.
 
 .. _v39-0-0: